Submit shell commands

Submit shell commands

bsub_cmd(cmd,
    name = NULL,
    hours = 1,
    memory = 1,
    cores = 1,
    temp_dir = bsub_opt$temp_dir,
    output_dir = bsub_opt$output_dir,
    dependency = NULL,
    enforce = bsub_opt$enforce,
    local = bsub_opt$local,
    sh_head = bsub_opt$sh_head,
    ...)

Arguments

cmd
A list of commands.

name
If name is not specified, an internal name calculated by digest is automatically assigned.

hours
Running time of the job.

memory
Memory usage of the job. It is measured in GB.

cores
Number of cores.

temp_dir
Path of temporary folder where the temporary R/bash scripts will be put.

output_dir
Path of output folder where the output/flag files will be put.

dependency
A vector of job IDs that current job depends on.

enforce
If a flag file for the job is found, whether to enforce to rerun the job.

local
Run job locally (not submitting to the LSF cluster)?

sh_head
Commands that are written as head of the sh script.

...
Command-line arguments can also be specified as name-value pairs.

Value

Job ID.

See also

Examples

if (FALSE) {
bsub_cmd("samtools sort ...", name = ..., memory = ..., cores = ..., ...)
}