Job dependencies

Job dependencies

job_dependency_all(job_tb = NULL)

job_dependency_igraph(job_id, job_tb = NULL)

job_dependency_dot(job_id, job_tb = NULL)

job_dependency_diagram(job_id, job_tb = NULL, ...)

Arguments

job_tb

A data frame from bjobs(). Internally used.

job_id

A single job ID.

...

Pass to DiagrammeR::grViz(), such as the size of the html widget.

Value

job_dependency_all() returns a list that contains three elements:

dep_mat: a two column matrix containing dependencies from parents to children.

id2name: a named vector containing mapping from job IDs to job names.

id2stat: a named vector containing mapping from job IDs to job status.

job_dependency_igraph() returns a igraph::igraph object which contains a dependency graph induced by the input job ID.

job_dependency_dot() returns a DOT code for GraphViz visualization.

job_dependency_diagram() makes a HTML-based dependency diagram.

Examples

if (FALSE) {
job1 = random_job()
job2 = random_job()
job3 = random_job(dependency = c(job1, job2))

job_dependency_all()
job_dependency_igraph(job3)
cat(job_dependency_dot(job3))
job_dependency_diagram(job3)
}